+++ R/AllClasses.R
@@ -1,18 +1,5 @@
 # CLASSES DEFINITION
 
-# Vector =======================================================================
-.BootstrapVector <- setClass(
-  Class = "BootstrapVector",
-  contains = "numeric"
-)
-.JackknifeVector <- setClass(
-  Class = "JackknifeVector",
-  slots = c(
-    hat = "numeric"
-  ),
-  contains = "numeric"
-)
-
 # MultivariateAnalysis =========================================================
 ## Results ---------------------------------------------------------------------
 #' Multivariate Data Analysis Results