+++ inst/examples/ex-bootstrap.R
@@ -1,15 +1,5 @@
 library(ggrepel)
 
-## Random samples from x with replacement
-x <- rnorm(20) # numeric
-boot <- bootstrap(x, do = mean, n = 100) # Sample mean
-summary(boot)
-
-## Sample observations from a multinomial distribution
-x <- sample(1:100, 100, TRUE) # integer
-boot <- bootstrap(x, do = median, n = 100)
-summary(boot)
-
 ## Partial bootstrap on CA
 ## Data from Lebart et al. 2006, p. 170-172
 color <- data.frame(
@@ -30,7 +20,7 @@ plot(X) +
 ## Bootstrap (30 replicates)
 Y <- bootstrap(X, n = 30)
 
-\donttest{
+\dontrun{
 ## Get replicated coordinates
 get_replications(Y, margin = 1)
 get_replications(Y, margin = 2)