+++ inst/examples/ex-jackknife.R
@@ -1,4 +1,0 @@
-## Jackknife
-x <- rnorm(20)
-jack <- jackknife(x, do = mean) # Sample mean
-summary(jack)